Use uint32 since code-generator don't support uint16
https://github.com/kubernetes/code-generator/blob/b6644f7753f281ae6b8ef94867f84b060e2f5e59/cmd/go-to-protobuf/protobuf/generator.go#L451

+++ cloud/upm.go
@@ -294,7 +294,7 @@ func isPatch(v1, v2 *semver.Version) bool {
 }
 
 // sortedSliceFromStringIntMap returns a slice of the keys in the map sorted alphabetically
-func sortedSliceFromStringIntMap(strMap map[string]uint16) []string {
+func sortedSliceFromStringIntMap(strMap map[string]uint32) []string {
 	strSlice := []string{}
 	for k := range strMap {
 		strSlice = append(strSlice, k)
